refactor(enhancement): split page view sections
