+++ src/components/enhancement/EnhancementView.tsx
@@ -1,0 +1,21 @@
+import React from 'react';
+import NavBar from '../NavBar';
+import EnhancementCharacterSection from './EnhancementCharacterSection';
+import EnhancementMaterialsSection from './EnhancementMaterialsSection';
+import EnhancementResultsSection from './EnhancementResultsSection';
+import EnhancementSettingsSection from './EnhancementSettingsSection';
+import type { EnhancementPageModel } from './useEnhancementPage';
+
+const EnhancementView: React.FC<{ model: EnhancementPageModel }> = ({ model }) => (
+  <div>
+    <NavBar />
+    <main className="enhancement-page max-w-4xl mx-auto px-4 py-6 space-y-5">
+      <EnhancementCharacterSection model={model} />
+      <EnhancementSettingsSection model={model} />
+      <EnhancementResultsSection model={model} />
+      <EnhancementMaterialsSection model={model} />
+    </main>
+  </div>
+);
+
+export default EnhancementView;
